+++ src/Symfony/Component/Notifier/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Add `AdminRecipientsProviderInterface`
  * Deprecate declaring `getAdminRecipients()` without implementing `AdminRecipientsProviderInterface`
+ * Deprecate reading a value that is not a boolean with `Dsn::getBooleanOption()`; it will throw in 9.0
 
 8.0
 ---